requirements.txt
# Core dependencies
python>=3.8
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pillow==10.1.0
python-multipart==0.0.6
python-docx==1.1.0
markdown==3.5.1

# ML and NLP
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
langchain==0.0.350
openai==1.6.0

# Vector database
chromadb==0.4.22

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# Object storage
minio==7.2.0

# Utilities
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
cachetools==5.3.2
psutil==5.9.6

# Async support
aiofiles==23.2.1
asyncio==3.4.3
httpx==0.25.2

# Monitoring and logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0